fix bug with localstorage
after reloading page (except homepage) the Cart is cleared and items from localstorage weren't parsed. It happened because when reload a page in Next.js, the server-side rendering (SSR) is triggered, which means the component's initial state is reinitialized. In this case, the initial state of cartItems is an empty array.

Solution: on each page call function updateCartFromLocalStorage in useEffect()